+++ .gitignore
@@ -68,3 +68,4 @@ dmypy.json
 
 
 .vscode
+*.db
